Feature: Store password hash in encrypted shared preferences
Previously, the database encryption password (used for login as well and created during the first launch) was directly stored encrypted in EncryptedSharedPreferences. With this commit, when a user inputs the password to create the user during the first launch, the password is hashed using PBKDF2WithHmacSHA256, and the resulting hash is stored in encrypted shared preferences.

Then, every time the user inputs the known password to login, the password is hashed using the same algorithm, the same number of iterations, and is confronted with the (hashed) password retrieved from the EncryptedSharedPreferences.

This improves security by ensuring that the actual password is not stored directly, but rather its hash, making it more difficult for unauthorized access to obtain sensitive information.